+++ src/main/java/duke/commands/AfterCommand.java
@@ -1,9 +1,10 @@
-package duke.Commands;
+package duke.commands;
 
-import duke.Common.DateTimeValidator;
-import duke.Common.Messages;
+import duke.common.DateTimeValidator;
+import duke.common.Messages;
 import duke.userinterface.Ui;
-import duke.tasks.*;
+import duke.tasks.TaskHelper;
+import duke.tasks.Storage;
 
 public class AfterCommand extends Command{
     public AfterCommand(String description) {
