+++ src/main/java/duke/commands/DeadlineCommand.java
@@ -1,18 +1,16 @@
-package duke.Commands;
+package duke.commands;
 
 import duke.tasks.Storage;
 import duke.tasks.TaskHelper;
 import duke.userinterface.Ui;
 
-public class ListCommand extends Command {
-
-    public ListCommand(String description) {
+public class DeadlineCommand extends Command{
+    public DeadlineCommand(String description) {
         super(description);
     }
 
     @Override
     public void execute(TaskHelper taskHelper, Ui ui, Storage storage) {
-        ui.printList(taskHelper.list());
+        ui.printToUser(taskHelper.deadline(description, storage));
     }
-
 }
