+++ src/main/java/duke/commands/DoneCommand.java
@@ -1,16 +1,16 @@
-package duke.Commands;
+package duke.commands;
 
 import duke.tasks.Storage;
 import duke.tasks.TaskHelper;
 import duke.userinterface.Ui;
 
-public class DeleteCommand extends Command {
-    public DeleteCommand(String description) {
+public class DoneCommand extends Command {
+    public DoneCommand(String description) {
         super(description);
     }
 
     @Override
     public void execute(TaskHelper taskHelper, Ui ui, Storage storage) {
-        ui.printToUser(taskHelper.delete(description, storage));
+        ui.printToUser(taskHelper.done(description, storage));
     }
 }
