+++ src/main/java/duke/commands/EventCommand.java
@@ -1,18 +1,16 @@
-package duke.Commands;
+package duke.commands;
 
 import duke.tasks.Storage;
 import duke.tasks.TaskHelper;
 import duke.userinterface.Ui;
 
-public class TodoCommand extends Command {
-
-    public TodoCommand(String description) {
+public class EventCommand extends Command{
+    public EventCommand(String description) {
         super(description);
     }
 
     @Override
     public void execute(TaskHelper taskHelper, Ui ui, Storage storage) {
-        ui.printToUser(taskHelper.todo(description, storage));
+        ui.printToUser(taskHelper.event(description, storage));
     }
-
 }
