+++ src/main/java/duke/commands/HelpCommand.java
@@ -1,18 +1,16 @@
-package duke.Commands;
+package duke.commands;
 
 import duke.tasks.Storage;
 import duke.tasks.TaskHelper;
 import duke.userinterface.Ui;
 
-public class FindCommand extends Command {
-
-    public FindCommand(String description) {
+public class HelpCommand extends Command{
+    public HelpCommand(String description) {
         super(description);
     }
 
     @Override
     public void execute(TaskHelper taskHelper, Ui ui, Storage storage) {
-        ui.printFiltered(taskHelper.filterKeyword(description), "match the keyword entered:");
+        ui.printHelp();
     }
-
 }
