+++ src/main/java/duke/commands/ListCommand.java
@@ -1,16 +1,18 @@
-package duke.Commands;
+package duke.commands;
 
 import duke.tasks.Storage;
 import duke.tasks.TaskHelper;
 import duke.userinterface.Ui;
 
-public class DeadlineCommand extends Command{
-    public DeadlineCommand(String description) {
+public class ListCommand extends Command {
+
+    public ListCommand(String description) {
         super(description);
     }
 
     @Override
     public void execute(TaskHelper taskHelper, Ui ui, Storage storage) {
-        ui.printToUser(taskHelper.deadline(description, storage));
+        ui.printList(taskHelper.list());
     }
+
 }
