+++ src/main/java/duke/tasks/Deadline.java
@@ -1,5 +1,5 @@
 package duke.tasks;
-import duke.Common.DateTimeValidator;
+import duke.common.DateTimeValidator;
 
 import java.time.LocalDateTime;
