+++ src/main/java/duke/tasks/Event.java
@@ -1,6 +1,6 @@
 package duke.tasks;
 
-import duke.Common.DateTimeValidator;
+import duke.common.DateTimeValidator;
 
 import java.time.LocalDateTime;
