+++ src/main/java/duke/tasks/TaskHelper.java
@@ -1,7 +1,7 @@
 package duke.tasks;
 
-import duke.Common.DateTimeValidator;
-import duke.Common.Messages;
+import duke.common.DateTimeValidator;
+import duke.common.Messages;
 import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
